requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
Brotli==1.0.9
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
cmake==3.27.4.1
dataclasses-json==0.5.14
dnspython==2.4.2
exceptiongroup==1.1.3
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.9.0
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.16.4
hyperframe==6.0.1
idna==2.10
Jinja2==3.1.2
langchain==0.0.287
langsmith==0.0.36
lit==16.0.6
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.6
numpy==1.25.2
packaging==23.1
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==3.12.0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rfc3986==1.5.0
safetensors==0.3.3
SQLAlchemy==2.0.20
sympy==1.12
tenacity==8.2.3
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.33.1
triton==2.0.0
typing-extensions==4.7.1
urllib3==2.0.4
wheel==0.38.4
yarl==1.9.2